requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
httptools==0.6.4
idna==3.10
loguru==0.7.3
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-onebot==2.4.6
nonebot2==2.4.1
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pygtrie==2.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.18.3
